added some features for loan part and some bugs fixed

+++ src/main/java/uz/com/model/dto/response/LoanResponse.java
@@ -18,7 +18,7 @@ public class LoanResponse {
 
     UUID id;
     BigDecimal amount;
-    Integer interestRate;
+    Double interestRate;
     LoanStatus status;
     LocalDate dueDate;
     UserResponse user;
